#29 Split UtilSeq class, Fasta and Genbank
